Refactor VirtualTile and remove legacy exception exports
Now the canonical place to import hw related exceptions is
iotile.core.hw.exceptions.  The legacy exports inside
iotile.core.hw.virtual have been removed.

Also removes async_tile_rpc decorator since that functionality has
been merged into tile_rpc and no longer needs a separate decorator
