+++ iotilecore/iotile/core/hw/virtual/__init__.py
@@ -6,17 +6,10 @@ from .virtualdevice_base import BaseVirtualDevice
 from .virtualdevice_standard import StandardVirtualDevice
 from .virtualdevice_simple import SimpleVirtualDevice
 
-from .common_types import (tile_rpc, RPCDispatcher, RPCInvalidIDError,
-                           RPCNotFoundError, RPCInvalidArgumentsError,
-                           RPCInvalidReturnValueError, TileNotFoundError,
-                           RPCErrorCode, BusyRPCResponse, unpack_rpc_payload, pack_rpc_payload,
-                           pack_rpc_response, unpack_rpc_response,
-                           VALID_RPC_EXCEPTIONS, rpc)
+from .common_types import (rpc, tile_rpc, unpack_rpc_payload, pack_rpc_payload,
+                           pack_rpc_response, unpack_rpc_response, RPCDispatcher)
 
-__all__ = ['BaseVirtualTile', 'VirtualTile', 'BaseVirtualDevice', 'StandardVirtualDevice', 'SimpleVirtualDevice', 'tile_rpc', 'rpc',
-           'RPCDispatcher', 'RPCInvalidIDError', 'TileNotFoundError',
-           'RPCNotFoundError', 'RPCInvalidArgumentsError', 'BusyRPCResponse',
-           'RPCInvalidReturnValueError', 'RPCErrorCode',
+__all__ = ['BaseVirtualTile', 'VirtualTile', 'BaseVirtualDevice',
+           'StandardVirtualDevice', 'SimpleVirtualDevice', 'tile_rpc', 'rpc',
            'unpack_rpc_payload', 'pack_rpc_payload', 'pack_rpc_response',
-           'unpack_rpc_response',
-           'VALID_RPC_EXCEPTIONS']
+           'unpack_rpc_response']
