+++ iotilecore/test/test_hw/test_virtualadapter.py
@@ -304,10 +304,11 @@ def test_recording_rpcs(tmpdir):
 
     rpc_lines = [",".join(x) for x in rpc_lines]
 
+    print(rpc_lines[4])
     assert rpc_lines == ['1,, 8,0x0004,0xc0,,                                        ,ffff74657374303101000003                ,',
                          '1,, 8,0x0004,0xc0,,                                        ,ffff74657374303101000003                ,',
                          '1,,11,0x0004,0xc0,,                                        ,ffff74657374303101000003                ,',
                          '1,,11,0x0004,0xc0,,                                        ,ffff74657374303101000003                ,',
-                         '1,, 8,0x8001,0xc0,,                                        ,01000000                                ,',
+                         '1,, 8,0x8001,0xc0,,                                        ,00000000                                ,',
                          '1,,11,0x8000,0xc0,,0300000005000000                        ,08000000                                ,',
-                         '1,,11,0x8001,0xc0,,                                        ,01000000                                ,']
+                         '1,,11,0x8001,0xc0,,                                        ,00000000                                ,']
