+++ iotileemulate/iotile/emulate/internal/__init__.py
@@ -5,7 +5,6 @@ main threads of tiles as well as dispatch rpcs between the tiles.
 """
 
 from .emulation_loop import EmulationLoop
-from .async_rpc import async_tile_rpc
 from .response import CrossThreadResponse, AwaitableResponse
 
-__all__ = ['EmulationLoop', 'async_tile_rpc', 'CrossThreadResponse', 'AwaitableResponse']
+__all__ = ['EmulationLoop', 'CrossThreadResponse', 'AwaitableResponse']
